LayoutTests/platform/chromium-mac/http/tests/xmlhttprequest/cross-origin-authorization-with-embedder-expected.txt

CONSOLE MESSAGE: line 18: Uncaught TypeError: Object [object Object] has no method 'authenticateSession'

Start
